had to fix because render gave build failed, issue was i was using useSearchParams() in page
